pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ilija <ilijapet@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
django-split-settings = "^1.2.0"
pyyaml = "^6.0.1"
django-environ = "^0.11.2"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
whitenoise = "^6.6.0"
pytest = "^7.4.4"
pytest-xdist = "^3.5.0"
pytest-django = "^4.7.0"
model-bakery = "^1.17.0"
django-debug-toolbar = "^4.2.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
colorlog = "^6.8.0"



[tool.isort]
multi_line_output = 5
line_length = 119


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "foundation_backend.project.settings"
django_find_project = false
python_files = "test_*.py"
testpaths = ["foundation_backend"]
filterwarnings = "ignore::DeprecationWarning:^(?!node\\.).*:"

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
